backend/requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
django-redis
django-celery-beat
aiohttp
psycopg2-binary
celery
redis
pytest
dj_database_url
python-dotenv
Pillow
pytest-django